refactor: 스케줄 인터페이스 제거
